sft_training/requirements.txt
unsloth
trl
transformers
datasets
accelerate
peft
bitsandbytes
huggingface_hub
python-dotenv
wandb
